Update interpretation rules for final and preliminary reports
* Also updated NLog due to a bug uncovering during unit testing

+++ HaemophilusWebTests/TestUtils/Attribute.cs
@@ -26,7 +26,10 @@ namespace HaemophilusWeb.TestUtils
 
         public override void AfterTest(ITest test)
         {
-            tempDirectoryToStoreTestData?.Delete(true);
+            if (tempDirectoryToStoreTestData.Exists)
+            {
+                tempDirectoryToStoreTestData?.Delete(true);
+            }
         }
 
         private void CreateTemporaryDirectoryAndAssignToTest(ITempDirectoryTest tempDirectoryTest)
